backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.40
aiosqlite==0.20.0
pydantic-settings==2.7.1
paho-mqtt==2.1.0
anthropic==0.43.0
httpx==0.28.1
praw==7.8.1
yt-dlp==2024.12.23
beautifulsoup4==4.12.3
numpy-stl==3.1.2
trimesh==4.5.3
rectpack==0.1
apscheduler==3.10.4
python-multipart==0.0.20
websockets==14.1
aiofiles==24.1.0
cryptography==44.0.0
